tests: add common.resource
common.resource is a Robot Framework `resource` file which, essentially,
is where one defines functions.

common/common.resource is a resource file, which contains
functions that are used by every application under tests/*.

Every directory under tests/ will have, inside it, a symlink to
the common/ directory.

When you run a RF test, you point to a directory (suite). RF copies this
suite to a tmp directory - thus meaning anything above the root of the
suite in the repository structure can't be referred to within the test
suite. Hence the symlink.

common.resource is starting out with these functions:
- EzClick
- Open Terminal
- Run Command in Terminal
- Close Terminal
- Install Debian Package
- Install Snap Package
- Start Application
- Log In